+++ src/factories/license.ts
@@ -19,45 +19,57 @@ Copyright (c) OWASP Foundation. All Rights Reserved.
 
 import type { DisjunctiveLicense, License } from '../models'
 import { LicenseExpression, NamedLicense, SpdxLicense } from '../models'
-import { fixupSpdxId } from '../spdx'
+import { fixupSpdxId, isValidSpdxLicenseExpression } from '../spdx'
 
 export class LicenseFactory {
   makeFromString (value: string): License {
     try {
+      return this.makeSpdxLicense(value)
+    } catch {
+      /* pass */
+    }
+
+    try {
       return this.makeExpression(value)
     } catch {
-      return this.makeDisjunctive(value)
+      /* pass */
     }
+
+    return this.makeNamedLicense(value)
   }
 
   /**
    * @throws {@link RangeError} if expression is not eligible
    */
-  makeExpression (value: string): LicenseExpression {
-    return new LicenseExpression(value)
+  makeExpression (value: string | any): LicenseExpression {
+    const expression = String(value)
+    if (isValidSpdxLicenseExpression(expression)) {
+      return new LicenseExpression(expression)
+    }
+    throw new RangeError('Invalid SPDX license expression')
   }
 
   makeDisjunctive (value: string): DisjunctiveLicense {
     try {
-      return this.makeDisjunctiveWithId(value)
+      return this.makeSpdxLicense(value)
     } catch {
-      return this.makeDisjunctiveWithName(value)
+      return this.makeNamedLicense(value)
     }
   }
 
   /**
    * @throws {@link RangeError} if value is not supported SPDX id
    */
-  makeDisjunctiveWithId (value: string | any): SpdxLicense {
-    const spdxId = fixupSpdxId(String(value))
-    if (undefined === spdxId) {
-      throw new RangeError('Unsupported SPDX id')
+  makeSpdxLicense (value: string | any): SpdxLicense {
+    const fixed = fixupSpdxId(String(value))
+    if (undefined === fixed) {
+      throw new RangeError('Unsupported SPDX license ID')
     }
 
-    return new SpdxLicense(spdxId)
+    return new SpdxLicense(fixed)
   }
 
-  makeDisjunctiveWithName (value: string | any): NamedLicense {
+  makeNamedLicense (value: string | any): NamedLicense {
     return new NamedLicense(String(value))
   }
 }