+++ webpack.config.js
@@ -56,16 +56,18 @@ const configBase = {
     }
   },
   externals: {
-    'packageurl-js': 'packageurl-js',
     ajv: 'ajv',
     'ajv-formats': 'ajv-formats',
-    'ajv-formats-draft2019': 'ajv-formats-draft2019'
+    'ajv-formats-draft2019': 'ajv-formats-draft2019',
+    'packageurl-js': 'packageurl-js',
+    'spdx-expression-parse': 'spdx-expression-parse'
   }
 }
 
 module.exports = [
   deepmerge(configBase, {
     mode: 'production',
+    devtool: 'source-map',
     output: {
       filename: 'lib.js'
     }